Fix the firing of "Jack"-Interrupts after resume when the modem is powered on.
GTA02_GPIO_nDL_GSM defaults to high/1. On resume do not enable the DL_GSM if
it was not enabled before. This is stopping the storm of interrupts.

Fix the logic in the download file handling. Downloads are disabled (0) when
the GTA02_GPIO_nDL_GSM is high (1). To enable downloading set
GTA02_GPIO_nDL_GSM to low (0, !on). Disable the jack interrupt while download
the is active. When disabling download we will get a couple of jack interrupts
but this is hardly avoidable.

Avoid reading the GPIO value if we do not even have a console set.

Signed-Off-by: Holger Freyther <zecke@openmoko.org>