change-backlight-level-not-forced-up-on-resume.patch
We at least always enabled backlight on resume, this patch
changes us to set backlight back to last requested backlight
brightness level on resume.  Note it means that you can
resume with screen blanked, but it should come back if that
happened with touchscreen action as usual.

/sys/class/backlight/pcf50633-bl/actual_brightness
and
/sys/class/backlight/pcf50633-bl/brightness

seem to agree after resume when reportedly they didn't before.

Signed-off-by: Andy Green <andy@openmoko.com>