debug-move-dev-info-to-dbg.patch Suggested-by: Sean McNeil <sean@mcneil.com>
To see if some subtle race is involved, Sean has tried
removing syslog traffic during resume and found he was
not seeing the resume crash any more.  We're giving it
a try to see if it changes the behaviour for anyone
else.  It would mean we have a pretty fine race in there
somewhere.

Signed-off-by: Andy Green <andy@openmoko.com>