[ARM]: Move asm/arch/gpio.h to mach/ directory
This file was moved in the big file move, but some OpenMoko specific
changes did not make it.  This patch peels out the relevant bits and
adds them to the gpio.h file in the upstream location.

The only OpenMoko specific change is the definition of gpio_to_irq and
irq_to_gpio.  These functions should really be defined in gpio_chip and
asm-generic/gpio.h; this is coming soon, but until then we'll just use
the Moko definitions that we've been using up until now.

This is not strictly correct for the GTA02 case, but it works given
the configuration that's currently in use.  This can be fixed (and
should become evident) when the configuration options are cleaned up.

Signed-off-by: Jonas Bonn <jonas.bonn@gmail.com>